requirements.txt
rpaframework==30.0.2     # https://rpaframework.org/releasenotes.html
robocorp==3.0.0           # https://pypi.org/project/robocorp
robocorp-browser==2.3.5   # https://pypi.org/project/robocorp-browser
requests==2.32.3          # https://pypi.org/project/requests
psycopg2-binary==2.9.9
python-dotenv
psycopg[binary]